+++ libraries/common/src/test/java/androidx/media3/common/FormatTest.java
@@ -20,10 +20,12 @@ import static androidx.media3.common.MimeTypes.VIDEO_MP4;
 import static androidx.media3.common.MimeTypes.VIDEO_WEBM;
 import static androidx.media3.test.utils.TestUtil.buildTestData;
 import static com.google.common.truth.Truth.assertThat;
+import static org.junit.Assert.assertThrows;
 
 import android.os.Bundle;
 import androidx.media3.test.utils.FakeMetadataEntry;
 import androidx.test.ext.junit.runners.AndroidJUnit4;
+import com.google.common.collect.ImmutableList;
 import java.util.ArrayList;
 import java.util.List;
 import org.junit.Test;
@@ -59,44 +61,58 @@ public final class FormatTest {
 
   @Test
   public void formatBuild_withLabelAndWithoutLabels_labelIsInLabels() {
-    Format format = createTestFormat();
-    format = format.buildUpon().setLabels(null).build();
+    Format format = new Format.Builder().setLabel("label").setLabels(ImmutableList.of()).build();
 
-    assertThat(format.labels.size()).isEqualTo(1);
-    assertThat(format.labels.get(0).value).isEqualTo(format.label);
+    assertThat(format.label).isEqualTo("label");
+    assertThat(format.labels).hasSize(1);
+    assertThat(format.labels.get(0).value).isEqualTo("label");
   }
 
   @Test
   public void formatBuild_withLabelsAndLanguageMatchingAndWithoutLabel_theLanguageMatchIsInLabel() {
-    Format format = createTestFormat();
-    format.labels.add(new Label("language", "matchingLabel"));
-    format = format.buildUpon().setLabel(null).build();
+    Format format =
+        new Format.Builder()
+            .setLabel(null)
+            .setLabels(
+                ImmutableList.of(
+                    new Label("en", "nonDefaultLabel"), new Label("zh", "matchingLabel")))
+            .setLanguage("zh")
+            .build();
 
     assertThat(format.label).isEqualTo("matchingLabel");
   }
 
   @Test
   public void formatBuild_withLabelsAndNoLanguageMatchingAndWithoutLabel_theFirstIsInLabel() {
-    Format format = createTestFormat();
-    format.labels.add(new Label("otherLanguage", "otherLabel"));
-    format = format.buildUpon().setLabel(null).build();
+    Format format =
+        new Format.Builder()
+            .setLabel(null)
+            .setLabels(
+                ImmutableList.of(new Label("fr", "firstLabel"), new Label("de", "secondLabel")))
+            .setLanguage("en")
+            .build();
 
-    assertThat(format.label).isEqualTo("label");
+    assertThat(format.label).isEqualTo("firstLabel");
   }
 
   @Test
-  public void formatBuild_withoutLabelsOrLabel_theyAreEmpty() {
+  public void formatBuild_withoutLabelsOrLabel_bothEmpty() {
     Format format = createTestFormat();
-    format = format.buildUpon().setLabel(null).setLabels(null).build();
+    format = format.buildUpon().setLabel(null).setLabels(ImmutableList.of()).build();
 
-    assertThat(format.label).isEqualTo(null);
-    assertThat(format.labels.size()).isEqualTo(0);
+    assertThat(format.label).isNull();
+    assertThat(format.labels).isEmpty();
   }
 
-  @Test(expected = IllegalStateException.class)
-  public void formatBuild_withLabelAndLabelsSetButNotMatching_throwsException() {
-    Format format = createTestFormat();
-    format.buildUpon().setLabel("otherLabel").build();
+  @Test
+  public void formatBuild_withLabelAndLabelsSetButNoMatch_throwsException() {
+    assertThrows(
+        IllegalStateException.class,
+        () ->
+            new Format.Builder()
+                .setLabel("otherLabel")
+                .setLabels(ImmutableList.of(new Label("en", "label")))
+                .build());
   }
 
   private static Format createTestFormat() {
@@ -125,12 +141,10 @@ public final class FormatTest {
             .setChromaBitdepth(11)
             .build();
 
-    List<Label> labels = new ArrayList<>();
-    labels.add(new Label("en", "label"));
     return new Format.Builder()
         .setId("id")
         .setLabel("label")
-        .setLabels(labels)
+        .setLabels(ImmutableList.of(new Label("en", "label")))
         .setLanguage("language")
         .setSelectionFlags(C.SELECTION_FLAG_DEFAULT)
         .setRoleFlags(C.ROLE_FLAG_MAIN)